Add files via upload
Job Posting Notifier
This repository contains a Python script designed to scrape the HTML contents of the StepStone webpage for job postings related to practical positions (Praktikum) in the field of computer science. The script identifies the latest job postings by searching for specific HTML classes associated with job advertisement div elements. Upon detecting a new job posting, the script interacts with the operating system to notify the user using desktop notifications.

Overview
The Job Posting Notifier script utilizes the BeautifulSoup library to parse HTML content and identify relevant job postings on the StepStone website. It then compares the current state of job postings with the previous state stored in a text file to determine if any new postings have been added. If a new posting is found, the script triggers a desktop notification using the Plyer library, informing the user about the availability of a new job opportunity.

Features
Automated web scraping of the StepStone website to extract job postings in the field of computer science.
Detection of new job postings by comparing the current and previous states of the webpage.
Desktop notification functionality to alert the user about new job opportunities.
Integration with the schedule library to run the scraping and notification process at regular intervals.